pyproject.toml
[tool.poetry]
name = "digital-ocean-fw-updater"
version = "0.1.0"
description = ""
authors = ["Zsolt Halo <net.zsolt.net@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
click = "^8.0.4"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fwupdate = "main:main"
